pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "stood-x"
version = "0.0.1"
authors = [
  { name="Ivan Sevillano-Garcia", email="isevillano@ugr.es" },
]
description = "STOOD-X methodology for OOD detection with explainability"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
        "zennit-crp[fast_img]>=0.6.0",
        "revel-xai>=1.0.3",
        "numpy>=2.1.2",
        "pytest>=8",
        "scikit-learn>=1.5.2",
        "scikit-image>=0.24.0",
        "torch>=2.4.1",
        "baycomp>=1.0.3",
        "matplotlib>=3.9.2",
        "scipy>=1.14.1",
        "pandas>=2.2.3",
        "torchvision>=0.19.1",
        "efficientnet_pytorch>=0.7.1",
        "tqdm>=4.66.5",
        "opencv-python>=4.10.0.84",
        "IPython>=8.28.0",
        "seaborn>=0.13.2",
        "plotly>=5.24.1",
        "ipywidgets>=8.1.5",
        "sphinx>=8.0.2",
        "sphinx_rtd_theme>=3.0.0",
        "sphinxcontrib.bibtex>=2.6.3",
        "nbsphinx>=0.9.5",
        "wget>=3.2",
        "pandoc>=2.4",
        "torchsummary>=1.5.1",
        "tensorboard>=2.18.0",
        "twine>=5.1.1",
        "statsmodels",
        "git+https://github.com/Jingkang50/OpenOOD",
        "libmr",

    ]
[tool.setuptools]
packages = ["STOODX"]

[project.urls]
Homepage = "https://github.com/ari-dasci/S-STOOD-X"